Renaming to specify CPU temperature sensor
