+++ src/bin/dual-iir.rs
@@ -194,7 +194,7 @@ mod app {
         dacs: (Dac0Output, Dac1Output),
         iir_state: [[iir::Vec5<f32>; IIR_CASCADE_LENGTH]; 2],
         generator: FrameGenerator,
-        temp_sensor: stabilizer::hardware::temp_sensor::CpuTempSensor,
+        cpu_temp_sensor: stabilizer::hardware::cpu_temp_sensor::CpuTempSensor,
     }
 
     #[init]
@@ -253,7 +253,7 @@ mod app {
             dacs: stabilizer.dacs,
             iir_state: [[[0.; 5]; IIR_CASCADE_LENGTH]; 2],
             generator,
-            temp_sensor: stabilizer.temperature_sensor,
+            cpu_temp_sensor: stabilizer.temperature_sensor,
         };
 
         // Enable ADC/DAC events
@@ -429,7 +429,7 @@ mod app {
         c.shared.network.lock(|net| net.direct_stream(target));
     }
 
-    #[task(priority = 1, shared=[network, settings, telemetry], local=[temp_sensor])]
+    #[task(priority = 1, shared=[network, settings, telemetry], local=[cpu_temp_sensor])]
     fn telemetry(mut c: telemetry::Context) {
         let telemetry: TelemetryBuffer =
             c.shared.telemetry.lock(|telemetry| *telemetry);
@@ -443,7 +443,7 @@ mod app {
             net.telemetry.publish(&telemetry.finalize(
                 gains[0],
                 gains[1],
-                c.local.temp_sensor.get_temperature().unwrap(),
+                c.local.cpu_temp_sensor.get_temperature().unwrap(),
             ))
         });
