+++ src/hardware/cpu_temp_sensor.rs
