+++ src/hardware/setup.rs
@@ -15,7 +15,7 @@ use smoltcp_nal::smoltcp;
 use super::{
     adc, afe, dac, design_parameters, eeprom, input_stamper::InputStamper,
     pounder, pounder::dds_output::DdsOutput, shared_adc::SharedAdc,
-    temp_sensor::CpuTempSensor, timers, DigitalInput0, DigitalInput1,
+    cpu_temp_sensor::CpuTempSensor, timers, DigitalInput0, DigitalInput1,
     EthernetPhy, NetworkStack, SystemTimer, Systick, AFE0, AFE1,
 };
